re-organise code
